refactor update method
