+++ services/app-api/utils/validation/schemaMap.ts
@@ -1,23 +1,190 @@
-import * as schema from "./schemas";
+import {
+  array,
+  mixed,
+  number as numberSchema,
+  string,
+  StringSchema,
+} from "yup";
 
+const error = {
+  REQUIRED_GENERIC: "A response is required",
+  REQUIRED_CHECKBOX: "Select at least one response",
+  INVALID_GENERIC: "Response must be valid",
+  INVALID_EMAIL: "Response must be a valid email address",
+  INVALID_URL: "Response must be a valid hyperlink/URL",
+  INVALID_DATE: "Response must be a valid date",
+  INVALID_END_DATE: "End date can't be before start date",
+  INVALID_NUMBER_OR_NA: 'Response must be a valid number or "N/A"',
+  INVALID_RATIO: "Response must be a valid ratio",
+};
+
+// TEXT
+export const text = (): StringSchema =>
+  string().typeError(error.INVALID_GENERIC).required(error.REQUIRED_GENERIC);
+export const textOptional = () => text().notRequired();
+
+// NUMBER - Helpers
+const validNAValues = ["N/A", "Data not available"];
+
+const ignoreCharsForSchema = (value: string, charsToReplace: RegExp) => {
+  return numberSchema().transform((_value) => {
+    return Number(value.replace(charsToReplace, ""));
+  });
+};
+
+// NUMBER - Number or Valid Strings
+export const number = () =>
+  mixed()
+    .test({
+      message: error.REQUIRED_GENERIC,
+      test: (val) => val != "",
+    })
+    .required(error.REQUIRED_GENERIC)
+    .test({
+      message: error.INVALID_NUMBER_OR_NA,
+      test: (val) => {
+        const replaceCharsRegex = /[,.]/g;
+        return (
+          ignoreCharsForSchema(val, replaceCharsRegex).isValidSync(val) ||
+          validNAValues.includes(val)
+        );
+      },
+    });
+export const numberOptional = () => number().notRequired();
+
+// Number - Ratio
+export const ratio = () =>
+  mixed()
+    .test({
+      message: error.REQUIRED_GENERIC,
+      test: (val) => val != "",
+    })
+    .required(error.REQUIRED_GENERIC)
+    .test({
+      message: error.INVALID_RATIO,
+      test: (val) => {
+        const replaceCharsRegex = /[,.:]/g;
+        const ratio = val.split(":");
+
+        // Double check and make sure that a ratio contains numbers on both sides
+        if (
+          ratio.length != 2 ||
+          ratio[0].trim().length == 0 ||
+          ratio[1].trim().length == 0
+        ) {
+          return false;
+        }
+
+        // Check if the left side of the ratio is a valid number
+        const firstTest = ignoreCharsForSchema(
+          ratio[0],
+          replaceCharsRegex
+        ).isValidSync(val);
+
+        // Check if the right side of the ratio is a valid number
+        const secondTest = ignoreCharsForSchema(
+          ratio[1],
+          replaceCharsRegex
+        ).isValidSync(val);
+
+        // If both sides are valid numbers, return true!
+        return firstTest && secondTest;
+      },
+    });
+
+// EMAIL
+export const email = () => text().email(error.INVALID_EMAIL);
+export const emailOptional = () => email().notRequired();
+
+// URL
+export const url = () => text().url(error.INVALID_URL);
+export const urlOptional = () => url().notRequired();
+
+// DATE
+export const date = () =>
+  string()
+    .required(error.REQUIRED_GENERIC)
+    .matches(dateFormatRegex, error.INVALID_DATE);
+export const dateOptional = () => date().notRequired();
+export const endDate = (startDateField: string) =>
+  date().test(
+    "is-after-start-date",
+    error.INVALID_END_DATE,
+    (endDateString, context) => {
+      const startDateString = context.parent[startDateField];
+      const startDate = new Date(startDateString);
+      const endDate = new Date(endDateString!);
+      return endDate >= startDate;
+    }
+  );
+
+// DROPDOWN
+export const dropdown = () =>
+  string().typeError(error.INVALID_GENERIC).required(error.REQUIRED_GENERIC);
+
+// CHECKBOX
+export const checkbox = () =>
+  array()
+    .min(1, error.REQUIRED_CHECKBOX)
+    .of(text())
+    .required(error.REQUIRED_CHECKBOX);
+export const checkboxOptional = () => checkbox().notRequired();
+export const checkboxSingle = () => array();
+
+// RADIO
+export const radio = () =>
+  array().min(1).of(text()).required(error.REQUIRED_GENERIC);
+export const radioOptional = () => radio().notRequired();
+
+// DYNAMIC
+export const dynamic = () =>
+  array().min(1).of(text()).required(error.REQUIRED_GENERIC);
+export const dynamicOptional = () => dynamic().notRequired();
+
+// NESTED
+export const nested = (
+  fieldSchema: Function | any,
+  parentFieldName: string,
+  parentOptionValue: any
+) => {
+  const fieldTypeMap = {
+    array: array(),
+    mixed: number(),
+    string: string(),
+    date: date(),
+  };
+  const fieldType: keyof typeof fieldTypeMap = fieldSchema().type;
+  const baseSchema: any = fieldTypeMap[fieldType];
+
+  return baseSchema.when(parentFieldName, {
+    is: (value: any) => value && value.indexOf(parentOptionValue) != -1,
+    then: () => fieldSchema(),
+  });
+};
+
+// REGEX
+export const dateFormatRegex =
+  /^((0[1-9]|1[0-2])\/(0[1-9]|1\d|2\d|3[01])\/(19|20)\d{2})|((0[1-9]|1[0-2])(0[1-9]|1\d|2\d|3[01])(19|20)\d{2})$/;
+
+// SCHEMA MAP
 export const schemaMap: any = {
-  text: schema.text(),
-  textOptional: schema.textOptional(),
-  number: schema.number(),
-  numberOptional: schema.numberOptional(),
-  ratio: schema.ratio(),
-  email: schema.email(),
-  emailOptional: schema.emailOptional(),
-  url: schema.url(),
-  urlOptional: schema.urlOptional(),
-  date: schema.date(),
-  dateOptional: schema.dateOptional(),
-  dropdown: schema.dropdown(),
-  checkbox: schema.checkbox(),
-  checkboxOptional: schema.checkboxOptional(),
-  checkboxSingle: schema.checkboxSingle(),
-  radio: schema.radio(),
-  radioOptional: schema.radioOptional(),
-  dynamic: schema.dynamic(),
-  dynamicOptional: schema.dynamicOptional(),
+  text: text(),
+  textOptional: textOptional(),
+  number: number(),
+  numberOptional: numberOptional(),
+  ratio: ratio(),
+  email: email(),
+  emailOptional: emailOptional(),
+  url: url(),
+  urlOptional: urlOptional(),
+  date: date(),
+  dateOptional: dateOptional(),
+  dropdown: dropdown(),
+  checkbox: checkbox(),
+  checkboxOptional: checkboxOptional(),
+  checkboxSingle: checkboxSingle(),
+  radio: radio(),
+  radioOptional: radioOptional(),
+  dynamic: dynamic(),
+  dynamicOptional: dynamicOptional(),
 };
